Add files via upload
creat playercard
The frontend application needs to be implemented using HTML, CSS, and JavaScript and
for data handling use local storage.You have to design the UI for all the web pages as
well.On the homepage show a card-based grid of all the teams. On the team card, the click
takes the user to the team details page.